+++ src/app/Components/PriceRange/constants.ts
@@ -1,4 +1,3 @@
 export const DEFAULT_RANGE = [0, 50000]
 export const RANGE_DOT_SIZE = 32
-export const SLIDER_STEP_VALUE = 100
 export const DEFAULT_PRICE_RANGE = "*-*"
