+++ src/app/Scenes/SavedSearchAlert/screens/AlertPriceRangeScreen.tsx
@@ -1,32 +1,37 @@
 import { Button, Flex, Spacer, Spinner, Text } from "@artsy/palette-mobile"
 import { StackScreenProps } from "@react-navigation/stack"
+import {
+  AlertPriceRangeScreenQuery,
+  AlertPriceRangeScreenQuery$data,
+} from "__generated__/AlertPriceRangeScreenQuery.graphql"
 import { PriceRange } from "app/Components/ArtworkFilter/Filters/helpers"
 import { SearchCriteria } from "app/Components/ArtworkFilter/SavedSearch/types"
 import { ArtworkFilterBackHeader } from "app/Components/ArtworkFilter/components/ArtworkFilterBackHeader"
 import { PriceRangeContainer } from "app/Components/PriceRange/PriceRangeContainer"
 import { DEFAULT_PRICE_RANGE } from "app/Components/PriceRange/constants"
-import { getBarsFromAggregations } from "app/Components/PriceRange/utils/getBarsFromAggregations"
+import { getBarsFromAggregations } from "app/Components/PriceRange/utils"
 import { CreateSavedSearchAlertNavigationStack } from "app/Scenes/SavedSearchAlert/SavedSearchAlertModel"
 import { SavedSearchStore } from "app/Scenes/SavedSearchAlert/SavedSearchStore"
 import { GlobalStore } from "app/store/GlobalStore"
-import { Suspense, useState } from "react"
+import { withSuspense } from "app/utils/hooks/withSuspense"
+import { useState } from "react"
 import { graphql, useLazyLoadQuery } from "react-relay"
 
-type AlertPriceRangeScreenProps = StackScreenProps<
+type AlertPriceRangeScreenQRProps = StackScreenProps<
   CreateSavedSearchAlertNavigationStack,
   "AlertPriceRange"
 >
 
-export const AlertPriceRangeScreen: React.FC<
-  StackScreenProps<CreateSavedSearchAlertNavigationStack>
-> = ({ navigation }) => {
-  const artistID = SavedSearchStore.useStoreState((state) => state.entity.artists[0].id)
-  const data = useLazyLoadQuery(AlertPriceRangeScreenQuery, {
-    artistID: artistID,
-  })
+interface AlertPriceRanceScreenProps extends AlertPriceRangeScreenQRProps {
+  artist: AlertPriceRangeScreenQuery$data["artist"]
+}
 
+export const AlertPriceRangeScreen: React.FC<AlertPriceRanceScreenProps> = ({
+  navigation,
+  artist,
+}) => {
   const histogramBars = getBarsFromAggregations(
-    (data as any).artist?.filterArtworksConnection?.aggregations
+    (artist as any)?.filterArtworksConnection?.aggregations
   )
   const attributes = SavedSearchStore.useStoreState((state) => state.attributes)
   const [filterPriceRange, setFilterPriceRange] = useState(
@@ -87,7 +92,7 @@ export const AlertPriceRangeScreen: React.FC<
   )
 }
 
-const AlertPriceRangeScreenQuery = graphql`
+const alertPriceRangeScreenQuery = graphql`
   query AlertPriceRangeScreenQuery($artistID: String!) {
     artist(id: $artistID) {
       filterArtworksConnection(aggregations: [SIMPLE_PRICE_HISTOGRAM], first: 0) {
@@ -104,10 +109,14 @@ const AlertPriceRangeScreenQuery = graphql`
   }
 `
 
-export const AlertPriceRangeScreenQueryRenderer: React.FC<AlertPriceRangeScreenProps> = (props) => {
-  return (
-    <Suspense fallback={<Spinner testID="alert-price-range-spinner" />}>
-      <AlertPriceRangeScreen {...props} />
-    </Suspense>
-  )
-}
+const Placeholder: React.FC<{}> = () => <Spinner testID="alert-price-range-spinner" />
+
+export const AlertPriceRangeScreenQueryRenderer: React.FC<AlertPriceRangeScreenQRProps> =
+  withSuspense((props) => {
+    const artistID = SavedSearchStore.useStoreState((state) => state.entity.artists[0].id)
+    const data = useLazyLoadQuery<AlertPriceRangeScreenQuery>(alertPriceRangeScreenQuery, {
+      artistID: artistID,
+    })
+
+    return <AlertPriceRangeScreen artist={data.artist!} {...props} />
+  }, Placeholder)
